refactor: extract format_time_ago to shared utils/time.rs
Move duplicate format_time_ago function from article_new.rs and
nostr_mention_dialog.rs into utils/time.rs as a shared utility.
Uses the more complete version that includes month formatting.
